placed static temp file creation method in file.cpp, created temp directory creation method in directory.cpp
